backend/requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
ruff==0.12.4
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
